opt-pulse/requirements.txt
fastapi
uvicorn
streamlit
polars
duckdb
pydantic
sqlalchemy
mysql-connector-python
openai
langchain-openai
pillow
python-dotenv